backend/requirements.txt
# Backend Python Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# LLM and RAG
langchain==0.1.0
langchain-community==0.0.13
openai==1.10.0

# Embeddings
sentence-transformers==2.3.1

# Vector Store
faiss-cpu==1.7.4

# Document Processing
pypdf==4.0.1
python-docx==1.1.0

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Utilities
